NeighborlyAPI_alb/createAdvertisements/__init__.py
import logging
import azure.functions as func
import pymongo

import sys
sys.path.insert(1, '../')

from db import MyDbConnection


def main(req: func.HttpRequest) -> func.HttpResponse:

    request = req.get_json()

    if request:
        try:
            url = MyDbConnection  # TODO: Update with appropriate MongoDB connection information
            client = pymongo.MongoClient(url)
            database = client['neighborly']
            collection = database['advertisements']

            rec_id1 = collection.insert_one(request)

            return func.HttpResponse(req.get_body())

        except ValueError:
            print("could not connect to mongodb")
            return func.HttpResponse('Could not connect to mongodb', status_code=500)

    else:
        return func.HttpResponse(
            "Please pass name in the body",
            status_code=400
        )